requirements.txt
streamlit>=1.35.0
gspread>=6.0.0
google-auth>=2.22.0
pandas>=2.0.0

pytz~=2025.2
firebase-admin
